moving radio button po into the component po and updating comments
